move faerie-related code from traps to output
